+++ plugins/control/control_command_attacher.py
@@ -8,10 +8,8 @@ class CommandAttacherPlugin(BasePlugin):
         """Forwards command with it's answer."""
         super().__init__()
 
-    async def global_before_message(self, msg, plugin):
+    async def global_before_message_checks(self, msg):
         if self.bot.settings.READ_OUT:
             return True
 
         msg.answer_values["forward_messages"] = msg.msg_id
-
-        return True